refactor: 학수번호 타입 수정
- Integer -> Long
- Wrapper 타입을 쓴 이유는 존재하지 않을 경우 null 값을 허용하기 위함

+++ src/main/java/com/gonghak98/v2/report/domain/abeek/basic/msc/MscBasic.java
@@ -12,7 +12,7 @@ import java.util.stream.Collectors;
 public class MscBasic implements Basic {
 
     private final Set<Course> essentialCourses;
-    private final Set<Integer> essentialCourseIds;
+    private final Set<Long> essentialCourseIds;
 
     public MscBasic(Set<Course> essentialCourses) {
         this.essentialCourses = essentialCourses;
@@ -24,7 +24,7 @@ public class MscBasic implements Basic {
     @Override
     public void checkAllCourses(List<CompletedCourse> completedCourses, RequirementResult requirementResult) {
         // MSC(Math, Science, Computing) 검사 로직 구현
-        Set<Integer> completedCourseIds = completedCourses.stream()
+        Set<Long> completedCourseIds = completedCourses.stream()
                                                           .map(CompletedCourse::getId)
                                                           .collect(Collectors.toSet());
