+++ src/test/java/com/gonghak98/v2/abeek/major/MajorTest.java
@@ -4,9 +4,9 @@ import static com.gonghak98.v2.abeek.fixture.MajorFixture.createGeneralMajor;
 import static com.gonghak98.v2.abeek.fixture.MajorFixture.createLabMajor;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.gonghak98.v2.abeek.fixture.GivenObjectFixture;
 import com.gonghak98.v2.report.domain.abeek.AreaType;
 import com.gonghak98.v2.report.domain.abeek.dto.RequirementResult;
-import com.gonghak98.v2.abeek.fixture.GivenObjectFixture;
 import com.gonghak98.v2.report.domain.abeek.major.Major;
 import com.gonghak98.v2.report.domain.student.CompletedCourse;
 import java.util.List;
@@ -23,7 +23,7 @@ public class MajorTest {
         @DisplayName("전공영역은 설계, 실험, 일반, 선후수 세부요건을 검사할 수 있다.")
         @CsvSource(value = {"5611, 디지털논리회로", "9658, 전기회로실험", "8076, 기초광학및실험", "-1, 전자소자공정실험", "9666, 마이크로컴퓨터실험"})
         @ParameterizedTest
-        void 전공영역_검사(int courseId, String studentLabCourseName) {
+        void 전공영역_검사(Long courseId, String studentLabCourseName) {
             //given
             CompletedCourse completedCourse = CompletedCourse.builder().id(courseId)
                                                              .name(studentLabCourseName)
